Feat/viz 50 api onair

+++ src/consumers/controller.consumer.ts
@@ -1,24 +1,36 @@
-import { ConsumerEvent, EventConsumer } from '../lib/events'
+import { ConsumerEvent, EventConsumer } from '../utilities/events'
 import { statSync, unlink, existsSync, mkdirSync } from 'fs'
-import { convertPath } from '../lib/utilities/filesystem'
+import { convertPath } from '../utilities/filesystem'
 import { dirname } from 'path'
-import { logger } from '../lib/utilities/logger'
-import { volume } from '../lib/utilities/filesystem'
+import { logger as baseLogger } from '../utilities/logger'
+import { volume } from '../utilities/filesystem'
+const logger = baseLogger.tag('Controller consumer')
 
 export interface IControllerConsumer {
   targetPath: string
   sourcePath: string
   retransferInterval?: bigint
   numberOfRetransfers?: bigint
+  onAirTimerInterval?: bigint
+}
+
+export interface IOnAirQueueObject {
+  eventName: string
+  eventData: any
 }
 
 export class ControllerConsumer extends EventConsumer {
   protected targetPath: string
   protected sourcePath: string
   protected retransferTimer: NodeJS.Timer
+  protected isOnAir: boolean
+  protected onAirEndTime: number
+  protected onAirTimer: NodeJS.Timer | null = null
+  protected onAirTimerInterval: BigInt
   protected retransferCache: {
     [key: string]: { data: { source: string; target: string }; attempts: number; emit: Function; emitted: boolean }
   }
+  protected onAirQueue: Array<any>
   protected numberOfRetransfers: number
 
   constructor(options: IControllerConsumer) {
@@ -27,18 +39,39 @@ export class ControllerConsumer extends EventConsumer {
     this.sourcePath = options.sourcePath
     this.validatePaths()
     this.retransferCache = {}
+    this.isOnAir = false
+    this.onAirEndTime = 0
+    this.onAirQueue = []
     this.retransferTimer = setInterval(() => this.retransfer(), Number(options.retransferInterval) || 10 * 1000)
     this.numberOfRetransfers = Number(options.numberOfRetransfers) || 3
+    this.onAirTimerInterval = options.onAirTimerInterval ?? 60000n
   }
 
   consume({ event, data, emit }: ConsumerEvent): void {
+    logger.data(data).trace(`Consumes event: ${event}`)
     switch (event) {
+      case 'on-air':
+        this.setOnAir({ event, data, emit })
+        break
+      case 'off-air':
+        this.setOffAir(emit)
+        break
       case 'file-changed':
+        if (this.isOnAir) {
+          this.addToOnAirQueue(event, data)
+        } else {
+          this.addFile({ event, data, emit })
+        }
+        break
       case 'file-added':
         this.addFile({ event, data, emit })
         break
       case 'file-deleted':
-        this.deleteFile({ event, data, emit })
+        if (this.isOnAir) {
+          this.addToOnAirQueue(event, data)
+        } else {
+          this.deleteFile({ event, data, emit })
+        }
         break
       case 'transfer-success':
         logger.info(`Transfered file ${data.target}`)
@@ -48,6 +81,53 @@ export class ControllerConsumer extends EventConsumer {
         logger.error(`Failed to transfer file ${data.target} due to ${data.stream} stream: ${data.error}`)
         break
     }
+  }
+
+  private addToOnAirQueue(eventName: string, eventData: any) {
+    const onAirQueueObject: IOnAirQueueObject = {
+      eventName: eventName,
+      eventData: eventData,
+    }
+    this.onAirQueue.push(onAirQueueObject)
+  }
+
+  private setOnAir({ data, emit }: ConsumerEvent) {
+    this.isOnAir = true
+    this.onAirEndTime = Number(data.endTime)
+    if (this.onAirTimer !== null) {
+      logger.trace('ClearInterval')
+      clearInterval(this.onAirTimer)
+    }
+    this.onAirTimer = setInterval(() => {
+      if (this.onAirEndTime > Date.now()) {
+        return
+      }
+      logger.trace('Calling setOffAir from onair timer')
+      this.setOffAir(emit)
+    }, Number(this.onAirTimerInterval))
+    logger.info(`Is now On Air.`)
+  }
+
+  private setOffAir(emit: any) {
+    if (!this.isOnAir) {
+      logger.debug('Already Off Air!')
+      return
+    }
+    this.isOnAir = false
+    logger.info(`Is now Off Air.`)
+
+    if (this.onAirTimer !== null) {
+      clearInterval(this.onAirTimer)
+      logger.trace('Cleared onAirTimer.')
+    }
+
+    logger.debug(`OnAirQueue length: ${this.onAirQueue.length}`)
+
+    this.onAirQueue.forEach((element) => {
+      logger.info(`Resending event: ${element.eventName} path: ${element.eventData.path}`)
+      emit(element.eventName, { path: element.eventData.path })
+    })
+    this.onAirQueue = []
   }
 
   private retransfer() {
@@ -60,7 +140,9 @@ export class ControllerConsumer extends EventConsumer {
   }
 
   private cacheForRetransfer(data: any, emit: any) {
-    if (data.target in this.retransferCache) {
+    const isInCache = data.target in this.retransferCache
+    logger.trace(`Is in cache: ${isInCache}`)
+    if (isInCache) {
       if (this.retransferCache[data.target].attempts >= this.numberOfRetransfers) {
         logger.warn(`Maximum retransfer tries reached for file: ${data.target}`)
         delete this.retransferCache[data.target]
@@ -79,11 +161,12 @@ export class ControllerConsumer extends EventConsumer {
   }
 
   private deleteFile({ data: { path } }: ConsumerEvent): void {
-    unlink(path.replace(this.sourcePath, this.targetPath), (err) => {
-      if (err) {
-        logger.error(err)
+    const filePath = path.replace(this.sourcePath, this.targetPath)
+    unlink(filePath, (error) => {
+      if (error) {
+        logger.data(error).error(`Failed to delete file: ${filePath}`)
       } else {
-        logger.info(`File deleted: ${path}`)
+        logger.info(`File deleted: ${filePath}`)
       }
     })
   }
@@ -106,8 +189,8 @@ export class ControllerConsumer extends EventConsumer {
           logger.debug('Transfer event sent: Source file has been updated')
           emit('transfer', { source: sourceFile, target: targetFile })
         }
-      } catch (err) {
-        logger.error(err)
+      } catch (error) {
+        logger.data({ source: sourceFile, target: targetFile, error }).error('Failed to transfer file')
       }
     } else {
       // Transfer file
